cocs/requirements.txt
playwright==1.45.0
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
websockets==12.0
pydantic==2.5.0
python-dotenv==1.0.0
asyncio-mqtt==0.16.1
loguru==0.7.2
httpx==0.25.2
psutil==5.9.8